REFACTOR: multiply and ADD: division implementing functional interfaces and recursion

+++ src/com/sofka/IDivide.java
@@ -3,6 +3,6 @@ package com.sofka;
 @FunctionalInterface
 public interface IDivide {
 
-    public Double calculate (Integer x, Integer y);
+    public Integer calculate (Integer x, Integer y);
 
 }
